feat: store result in a different file to make cwl happy
